[APG-959] Rename Find And Refer API to Find And Refer Intervention API across the codebase and update related configuration
