+++ helm_deploy/hmpps-accredited-programmes-manage-and-deliver-api/values.yaml
@@ -38,8 +38,8 @@ generic-service:
     hmpps-accredited-programmes-manage-and-deliver-api-application-insights:
       APPLICATIONINSIGHTS_CONNECTION_STRING: "APPLICATIONINSIGHTS_CONNECTION_STRING"
 
-    SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_FIND-AND-REFER-API_CLIENT-ID: "SYSTEM_CLIENT_ID"
-    SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_FIND-AND-REFER-API_CLIENT-SECRET: "SYSTEM_CLIENT_SECRET"
+    SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_FIND-AND-REFER-INTERVENTION-API_CLIENT-ID: "SYSTEM_CLIENT_ID"
+    SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_FIND-AND-REFER-INTERVENTION-API_CLIENT-SECRET: "SYSTEM_CLIENT_SECRET"
 
     rds-postgresql-instance-output:
       DB_HOST: "rds_instance_endpoint"
