+++ helm_deploy/values-dev.yaml
@@ -12,7 +12,7 @@ generic-service:
     SPRING_PROFILES_ACTIVE: "dev"
     APPLICATIONINSIGHTS_CONFIGURATION_FILE: "applicationinsights.dev.json"
     HMPPS_AUTH_URL: "https://sign-in-dev.hmpps.service.justice.gov.uk/auth"
-    SERVICES_FIND-AND-REFER-API_BASE-URL: https://prison-api-dev.prison.service.justice.gov.uk
+    SERVICES_FIND-AND-REFER-INTERVENTION-API_BASE-URL: "https://find-and-refer-intervention-api-dev.hmpps.service.justice.gov.uk"
 
 
 # CloudPlatform AlertManager receiver to route prometheus alerts to slack
