+++ src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/client/findAndReferInterventionApi/FindAndReferInterventionApiClient.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.client.findAndReferApi
+package uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.client.findAndReferInterventionApi
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.slf4j.LoggerFactory
@@ -6,12 +6,12 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.client.BaseHMPPSClient
-import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.client.findAndReferApi.model.ReferralDetails
+import uk.gov.justice.digital.hmpps.accreditedprogrammesmanageanddeliverapi.client.findAndReferInterventionApi.model.ReferralDetails
 import java.util.UUID
 
 @Component
-class FindAndReferApiClient(
-  @Qualifier("findAndReferApiWebClient") webClient: WebClient,
+class FindAndReferInterventionApiClient(
+  @Qualifier("findAndReferInterventionApiWebClient") webClient: WebClient,
   objectMapper: ObjectMapper,
 ) : BaseHMPPSClient(webClient, objectMapper) {
 
@@ -19,8 +19,7 @@ class FindAndReferApiClient(
     private val log = LoggerFactory.getLogger(this::class.java)
   }
 
-  fun getReferral(referralId: UUID) = getRequest<ReferralDetails>("Find And Refer API") {
-    log.debug("Retrieving referral details for referral id: $referralId")
+  fun getReferral(referralId: UUID) = getRequest<ReferralDetails>("Find And Refer Intervention API") {
     path = "/referral/$referralId"
   }
 }
